requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
blessings==1.6
bpython==0.16
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
click==6.7
curtsies==0.2.11
decorator==4.0.11
Flask==0.12.2
git-review==1.25.0
greenlet==0.4.12
idna==2.5
ipdb==0.10.3
ipython==5.3.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.0.2
MarkupSafe==1.0
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
pyxdg==0.25
requests==2.17.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.21.1
vine==1.1.3
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.12.2
